Don't use I2C mock in deployed code.
